Move TestAnalyzer to dedicated file
